==> 00_DF_prep.R <==

library(easycsv)
directory = "000_DFs/"
loadcsv_multi(directory, extension = "CSV")


abbvie$Company <- "AbbVie"

adobe$Company <- "Adobe"

Aflac$Company <- "Aflac"

Airbnb$Company <- "Airbnb"

AlaskaAirlines$Company <- "Alaska Airlines"

Albertsons$Company <- "Albertsons Companies"

Allergan$Company <- "Allergan"

alorica$Company <- "Alorica"

amazon$Company <- "Amazon"

Amc$Company <- "AMC Entertainment"

Amedisys$Company <- "Amedisys"

americanAirlines$Company <- "American Airlines"

ann_inc$Company <- "Ann Taylor"

Anthem$Company <- "Anthem"

arbys$Company <- "Arby's"

ascension$Company <- "Ascension"

AstraZeneca$Company <- "AstraZeneca"

Atos$Company <- "Atos"

atrium$Company <- "Atrium Health"

att$Company <- "AT&T"

Auto_Zone$Company <- "AutoZone"

Autodesk$Company <- "Autodesk"

AvisBudget$Company <- "Avis Budget Group"

BainCompany$Company <- "Bain & Company"

BananaRep$Company <- "Banana Republic"

BarnesNoble$Company <- "Barnes & Noble"

Bayer$Company <- "Bayer"

BaylorSW$Company <- "Baylor Scott & White Health"

BedBathBeyond$Company <- "Bed Bath & Beyond"

BestBuy$Company <- "Best Buy"

BlackRock$Company <- "BlackRock"

Bloomingdale$Company <- "Bloomingdale's"

BristolMyers$Company <- "Bristol Myers Squibb"

BP$Company <- "BP"

BrighamWH$Company <- "Brigham Womans Hospital"

Burlington$Company <- "Burlington Stores"

CACI$Company <- "CACI"

CapriHoldings$Company <- "Capri Holdings"

Carnival$Company <- "Carnival"

Carter$Company <- "Carter's"

CAtechnologies$Company <- "CA Technologies"

Caterpillar$Company <- "Caterpillar"

Centene$Company <- "Centene"

Ceridian$Company <- "Ceridian"

CGI$Company <- "CGI"

Chase$Company <- "Chase"

ChaseSchwab$Company <- "Charles Schwab"

Chevron$Company <- "Chevron"

ChoiceHotels$Company <- "Choice Hotels"

CHS$Company <- "CHS Inc"

Citrix$Company <- "Citrix"

ClevelandClinic$Company <- "Cleveland Clinic"

Clorox$Company <- "Clorox"

Cognizant$Company <- "Cognizant Technology Solutions"

Comerica$Company <- "Comerica"

CompuCom$Company <- "CompuCom"

DairyQueen$Company <- "Dairy Queen"

Darden$Company <- "Darden"

Dell$Company <- "Dell"

Dennys$Company <- "Denny's"

Discover$Company <- "Discover"

DiscoveryInc$Company <- "Discovery Inc"

Dominos$Company <- "Domino's"

DunkinDonuts$Company <- "Dunkin' Donuts"

DuPont$Company <- "DuPont"

Eaton$Company <- "Eaton"

EdwardJones$Company <- "Edward Jones"

ElectronicArts$Company <- "Electronic Arts"

Enterprise$Company <- "Enterprise"

Epic$Company <- "Epic"

ExpressScripts$Company <- "Express Scripts"

Facebook$Company <- "Facebook"

FactSet$Company <- "FactSet"

FiatCA$Company <- "Fiat Chrysler Automobiles"

FidelityInv$Company <- "Fidelity Investments"

FifthThird$Company <- "Fifth Third"

Fluor$Company <- "Fluor"

Fossil$Company <- "Fossil"

FreseniusMC$Company <- "Fresenius"

FrontierComm$Company <- "Frontier Communications"

GEhealthCare$Company <- "GE Healthcare"

Geico$Company <- "Geico"

GeneralMills$Company <- "General Mills"

GiantFood$Company <- "Giant Food"

GlaxoSmithKline$Company <- "GlaxoSmithKline"

GoldmanSachs$Company <- "Goldman Sachs"

Goodyear$Company <- "Goodyear"

Google$Company <- "Google"

Grubhub$Company <- "Grubhub"

Guess$Company <- "Guess"

Halliburton$Company <- "Halliburton"

HarrisTeeter$Company <- "Harris Teeter"

HCLtech$Company <- "HCL Tech"

HCSC$Company <- "HCSC"

Henkel$Company <- "Henkel"

HnM$Company <- "H&M"

HomeDepot$Company <- "The Home Depot"

HSBC$Company <- "HSBC Holdings"

Hyatt$Company <- "Hyatt"

HyVee$Company <- "Hy-Vee"

Infosys$Company <- "Infosys"

InNourBurger$Company <- "In-N-Out Burger"

Inovalon$Company <- "Inovalon"

IntermountainHC$Company <- "Intermountain"

JackintheBox$Company <- "Jack in the Box"

Jacobs$Company <- "Jacobs"

JetBlue$Company <- "JetBlue"

JPMorgan$Company <- "JPMorgan"

JudeCH$Company <- "St. Jude Children's Hospital"

Kellogg$Company <- "Kellogg Compnay"

KFC$Company <- "KFC"

KiewitCorp$Company <- "Kiewit Corporation"

KimberlyClark$Company <- "Kimberly-Clark"

KPMG$Company <- "KPMG"

Lenovo$Company <- "Lenovo"

LinkedIn$Company <- "LinkedIn"

LockheedMartin$Company <- "Lockheed Martin"

Lululemon$Company <- "Lululemon"

Lyft$Company <- "Lyft"

MagellanHealth$Company <- "Magellan Health"

MarriottInt$Company <- "Marriott International"

Marshalls$Company <- "Marshalls"

MathWorks$Company <- "MathWorks"

MayoClinic$Company <- "Mayo Clinic"

McDonalds$Company <- "McDonald's"

McKinsey$Company <- "McKinsey & Company"

MemorialSK$Company <- "Memorial Sloan Kettering"

Mercer$Company <- "Mercer"

MerrillLynch$Company <- "Merrill Lynch"

microsoft$Company <- "Microsoft"

MondelezInt$Company <- "Mondelez International"

MorganStanley$Company <- "Morgan Stanley"

Nationwide$Company <- "Nationwide"

NatOilVar$Company <- "National Oilwell Varco"

NBCUniversal$Company <- "NBCUniversal"

Nestle$Company <- "Nestle"

Netflix$Company <- "Netflix"

NorthernTrust$Company <- "Northern Trust"

NorthwesternMu$Company <- "Northwestern Mutual"

NYlife$Company <- "New York Life"

NYpresbyterianHost$Company <- "NewYork-Presbyterian Hospital"

OReillyAutoPArts$Company <- "O'Reilly Auto Parts"

Panasonic$Company <- "Panasonic"

PandaExp$Company <- "Panda Express"

Panera$Company <- "Panera Bread"

ParkerHannifin$Company <- "Parker Hannifin"

Paycom$Company <- "Paycom"

Paylocity$Company <- "Paylocity"

PitneyBowes$Company <- "Pitney Bowes"

PnG$Company <- "Procter & Gamble"

PPG$Company <- "PPG"

PwC$Company <- "PwC"

RalfLAuren$Company <- "Ralph Lauren"

RaytheonTech$Company <- "Raytheon Technologies"

RBC$Company <- "RBC"

RedRobin$Company <- "Red Robin"

RegeneronPharma$Company <- "Regeneron Pharmaceuticals"

ResCare$Company <- "ResCare"

RiteAid$Company <- "Rite Aid"

Rockwell$Company <- "Rockwell Automation"

Safeway$Company <- "Safeway"

SAIC$Company <- "SAIC"

Sanofi$Company <- "Sanofi"

SaveAlot$Company <- "Save-A-Lot Food Stores"

SeaWorld$Company <- "SeaWorld Parks & Entertainment"

ServiceSource$Company <- "ServiceSource"

Shell$Company <- "Shell"

SkyWestAirlines$Company <- "SkyWest Airlines"

Slalom$Company <- "Slalom"

SpaceX$Company <- "SpaceX" 

Starbucks$Company <- "Starbucks"

StateFarm$Company <- "State Farm"

SteakShake$Company <- "Steak n Shake"

StitchFix$Company <- "Stitch Fix"

StopShop$Company <- "Stop & Shop"

Sutherland$Company <- "Sutherland"

TacoBell$Company <- "Taco Bell"

TakedaPharma$Company <- "Takeda Pharmaceuticals"

TD$Company <- "TD"

TenetHealth$Company <- "Tenet Health"

Tesla$Company <- "Tesla"

TexasRoadHouse$Company <- "Texas Roadhouse"

TIAA$Company <- "TIAA"

TIauto$Company <- "TI Automotive"

TractorSupply$Company <- "Tractor Supply"

TrilogyHS$Company <- "Trilogy Health Services"

TRowePrice$Company <- "T. Rowe Price"

UltaBeauty$Company <- "Ulta Beauty"

Unilever$Company <- "Unilever"

UnitedHealth$Company <- "UnitedHealth Group"

UPMC$Company <- "UPMC"

UrbanOutfitters$Company <- "Urban Outfitters"

USAA$Company <- "USAA"

USbank$Company <- "U.S. Bank"

Verizon$Company <- "Verizon"

VitasHC$Company <- "Vitas Healthcare"

VMware$Company <- "VMware"

Walmart$Company <- "Walmart"

WarnerBros$Company <- "Warner Brothers"

Weatherford$Company <- "Weatherford"

Whataburger$Company <- "Whataburger"

WhirlpoolCorp$Company <- "Whirlpool Corporation"

WholeFoodsMarket$Company <- "Whole Foods Market"

WyndhamDest$Company <- "Wyndham Destinations"

xpo$Company <- "XPO Logistics"

yahoo$Company <- "Yahoo"

zillow$Company <- "Zillow"

#objects()

library(tidyverse)
d <- bind_rows(abbvie,adobe,Aflac,Airbnb,AlaskaAirlines,Albertsons,Allergan,alorica,amazon,Amc,Amedisys,americanAirlines,ann_inc,Anthem,
               arbys,ascension,AstraZeneca,Atos,atrium,att,Auto_Zone,Autodesk,AvisBudget,
               BainCompany,BananaRep,BarnesNoble,Bayer,BaylorSW,BedBathBeyond,BestBuy,BlackRock,Bloomingdale,BristolMyers,BP,BrighamWH,Burlington,
               CACI,CapriHoldings,Carnival,Carter,CAtechnologies,Caterpillar,Centene,Ceridian,CGI,Chase,ChaseSchwab,Chevron,ChoiceHotels,
               CHS,Citrix,ClevelandClinic,Clorox,Cognizant,Comerica,CompuCom,
               DairyQueen,Darden,Dell,Dennys,Discover,DiscoveryInc,Dominos,DunkinDonuts,DuPont,Eaton,EdwardJones,ElectronicArts,Enterprise,
               Epic,ExpressScripts,Facebook,FactSet,FiatCA,FidelityInv,FifthThird,Fluor,Fossil,
               FreseniusMC,FrontierComm,GEhealthCare,Geico,GeneralMills,GiantFood,GlaxoSmithKline,GoldmanSachs,Goodyear,Google,Grubhub,
               Guess,Halliburton,HarrisTeeter,HCLtech,HCSC,Henkel,HnM,HomeDepot,HSBC,
               Hyatt,HyVee,Infosys,InNourBurger,Inovalon,IntermountainHC,JackintheBox,Jacobs,JetBlue,JPMorgan,JudeCH,Kellogg,KFC,
               KiewitCorp,KimberlyClark,KPMG,Lenovo,LinkedIn,LockheedMartin,Lululemon,Lyft,
               MagellanHealth,MarriottInt,Marshalls,MathWorks,MayoClinic,McDonalds,McKinsey,MemorialSK,Mercer,MerrillLynch,
               microsoft,MondelezInt,MorganStanley,Nationwide,NatOilVar,NBCUniversal,Nestle,Netflix,NorthernTrust,
               NorthwesternMu,NYlife,NYpresbyterianHost,OReillyAutoPArts,Panasonic,PandaExp,Panera,ParkerHannifin,Paycom,Paylocity,
               PitneyBowes,PnG,PPG,PwC,RalfLAuren,RaytheonTech,RBC,RedRobin,RegeneronPharma,ResCare,
               RiteAid,Rockwell,Safeway,SAIC,Sanofi,SaveAlot,SeaWorld,ServiceSource,Shell,SkyWestAirlines,Slalom,SpaceX,Starbucks,StateFarm,
               SteakShake,StitchFix,StopShop,Sutherland,TacoBell,TakedaPharma,TD,TenetHealth,Tesla,
               TexasRoadHouse,TIAA,TIauto,TractorSupply,TrilogyHS,TRowePrice,UltaBeauty,Unilever,UnitedHealth,UPMC,
               UrbanOutfitters,USAA,USbank,Verizon,VitasHC,VMware,Walmart,WarnerBros,Weatherford,Whataburger,WhirlpoolCorp,
               WholeFoodsMarket,WyndhamDest,xpo,yahoo,zillow)

library(readxl)
comp_with_ind <- read_excel("000_data/company_list_ss_with_industry.xlsx")

dj <- left_join(d, comp_with_ind, by = "Company")

rm(list=setdiff(ls(), "dj")); d <- dj; rm(dj)

d$rowID <- 1:nrow(d)
library(stringi)
set.seed(666)
nrow(d)
d$strID <- stri_rand_strings(n = nrow(d), length = 6)
d$EmCoID <- "EM"

d <- d %>% select(23,21,22,1:20)
  
library(janitor)

d %>% tabyl(Company)

# d %>% select(Company) %>% ggplot(., aes(x=Company)) +
#   geom_bar(aes(y = (..count..)),fill="grey", alpha=0.6) +
#   geom_text(aes(y = (..count..) ,label =   ifelse((..count..)==0,"",scales::percent((..count..)/sum(..count..)))), stat="count",colour="black") +
#   theme_bw() +
#   xlab("") +
#   ylab("Frequency") +
#   labs(title="Company") +
#   coord_flip()

d %>% tabyl(Industry)

d %>% select(Industry) %>% ggplot(., aes(x=Industry)) +
  geom_bar(aes(y = (..count..)),fill="grey", alpha=0.6) +
  geom_text(aes(y = (..count..) ,label =   ifelse((..count..)==0,"",scales::percent((..count..)/sum(..count..)))), stat="count",colour="black") +
  theme_bw() +
  xlab("") +
  ylab("Frequency") +
  labs(title="Industry") +
  coord_flip()

d <- d %>% mutate_if(is.character, list(~na_if(.,"")))

save(d, file = "000_data/raw_coprus_DF.RData")
